requirements.txt
pyautogen
python-dotenv
torch
numpy
spacy
requests
git+https://github.com/openai/whisper.git
groq
sounddevice
soundfile
gradio